refactor: make AccessAnalyzer envoy-specific for now
